unit tests for migrate_passwd_file_except_root()
This function looks like a possible place in which to
implement issue #36[1], consistent UID/GID allocation.
With that in mind, this commit implements a small set
of unit tests to act as a starting point for said work.

The work won't be easy: the calling interface needs to
be changed, either by adding an ostree (against which
to compare) or a JSON pointer to config settings with
an enumeration of UIDs and GIDs. Or maybe there's a
better way. Regardless, with this test, such work
can proceed. Or even if not, at least we have tests
and a skeleton for building more.

Note that this uses the check[2] framework, and requires:

    # yum install check check-devel

There's probably a way to tell configure.ac to proceed
even if that's missing; I'll leave that to autoconf experts.

 [1] https://github.com/projectatomic/rpm-ostree/issues/36
 [2] http://check.sourceforge.net/doc/check_html/index.html#Top